enh: moved contiguous statement to depend on BUD_FORTRAN>=2003
The contiguous statement is dependent on the 2003 specification.
Hence, if the fortran version is denoted sub-2003, we do not add
contiguous statements.
However, it may be added if needed using:
   -DBUD_FORTRAN_CONTIGUOUS=", contiguous"

A small change in the stringification of the names have been made
to go beyond the pre-processor removing C-comments (//)
This means that one may not use // as concatenation because
CPP removes those statements.

Signed-off-by: Nick Papior <nickpapior@gmail.com>