enh: Updated naming routines and general updates
All delete_data are changed to common_delete_.
This ensures that other buds may contain the routine
delete_data.

Added copy routine for actual copy of data.
This also requires a common_copy_ routine
which copies the TYPE_ data structures. Currently
it only copies the error_ variable as the ID shouldn't
be the same, nor the refs_.

Signed-off-by: Nick Papior <nickpapior@gmail.com>