enh: added MP_Comm and updated Dist1D
Added a MP_Comm module which is a wrapper for MPI.
It contains a reference counted data type which holds information
regarding a communicator, its group, the rank and size of the
communicator.
It wraps all of the most used routines in both OO and non-OO routines
without prefixed routine names.
It also wraps some of the communicator splits into generic function
calls such that the interface decides the call:
i.e.
 call Comm_Create
can refer to:
 MPI_Comm_Create
or
 MPI_Comm_Create_Group
depending on the arguments.

Updated Dist1D to contain the MP_Comm type. This off-loads many routines
to the MP_Comm type and cleans the Dist object.

Enabled building only the sources by calling:
 make source
one, may additionally do:
 make source-src
or
 make source-mpi
to construct the sources in the sub-directories.

Signed-off-by: Nick Papior <nickpapior@gmail.com>