api: moved SP to SM
It was not very informative to call the sparse matrices for
patterns. So we have, again, reverted the data-structures.

Also, all common modules should be suffixed with _common
for consistency. This is now the standard for SM, Lists etc.

Enabled a single module to use all precisisons for a majority
of the data-types.
TODO: should we create a single module which encapsulates
EVERYTHING?

Signed-off-by: Nick Papior <nickpapior@gmail.com>