bug: fixed reference counts when using function retrieval
When retrieving elements in reference counted objects the
reference counts is intrinsic for correct behaviour.
For instance, if one overloads the = operator, one
may easily do:
  out = this
while if you use a function for retrieval it actually
does two increments if one uses assignment in element
function:
  out = element(this)
hence the function should use pointer assignment
in the function retrieval to ensure a correct reference
count.

Signed-off-by: Nick Papior <nickpapior@gmail.com>