enh: updated API for array creations, shape and dimension retrieval
Instead of doing array copy as a routine for creation it might
be better to pass a shape which may be used to create the array.
This, sadly, means that one cannot also have a copy creation
routine interfaced to new() because when the precision is
integer there will be two equivalent interfaces.
So, we prefer the shape creation.

Also added dimension(this) function to retrieve the
dimensionality of the array. Although there is no ambiguity
because of the type designation it may be nice when dealing
with preprocessed code.

This has also been added to the SM_Array.inc code.

Wrapped some retrieval functions in the Array.inc and
SM* modules to check for is_initd(this) to ensure
correctness.

Signed-off-by: Nick Papior <nickpapior@gmail.com>