enh: communicator additions and fixes
Added Wait/Test functions to the communicator.

Wrapped retrieval functions with is_initd(this)
to always return something.

When deleting a communicator it stalled when
using MPI_Comm_Disconnect. It should only stall
if there are non-closed communications on the network.
But I do not think there were. So currently I change
it to MPI_Comm_Free, in this case the user should
be adviced to call Barrier() just before doing a
delete to ensure completed communications (unless
there are non-blocking communications).

Signed-off-by: Nick Papior <nickpapior@gmail.com>