enh: added distribute function for Dist1D_SM_Array
Now a given distributed matrix can very easily be distributed
using a single call to a function. It currently can deal
with all, but the user distribution and is quite efficient
due to non-blocking, non-buffered sends. The recieving end
is here the limitation because it uses blocking recv.
This may easily be relieved in the data array messages because
there we know the size of the messages.
However, it requires additional request arrays.

Signed-off-by: Nick Papior <nickpapior@gmail.com>