enh: added association routines for transfering N-D to M-D
The routines are currently existing in a transfering of a
pointer or via array copying.
This makes it very easy to use simple 1D arrays in many routines
which may then be casted to more conforming arrays when needed.
An example could be:
   real :: in(100,10)
   real :: out(10,10,10)
   call dim_transfer(10,10,10,out,100,10,in)
which copies the first 1000 elements from in into the first
1000 elements of out.

Signed-off-by: Nick Papior <nickpapior@gmail.com>