enh: Ensured all buds to have a read/write function
A read/write function is now inherent to bud_common.inc
This has the consequence that these routines are now forced:
 - new_
 - delete_
 - print_ (unless BUD_NO_PRINT)
 - copy_ (unless BUD_NO_COPY)
 - read_/write_ (unless BUD_NO_IO)

The Collections object now tries to read/write all containers.
In some cases the collections need a separate write/read. In this case
the definition of BUD_IO_OVERLOAD ensures the public interfaces read/write
but does not implement read_/write_.
Currently we lack a way to say that only some of the collections have a read/write
method.

The FiniteStack object also implements the read/write of all existing elements.

Signed-off-by: Nick Papior <nickpapior@gmail.com>