enh: New error handler
Now all errors are stored in the top level bud.
I.e. not in the BUD_ type. The reason is that in some cases
it was necessary to return an error value, say if input
parameters where wrong and a new bud could not be allocated.
In this case one could not return an error message without
allocating the object.
By having the error message associated with the BUD type
we can _always_ return an error message.

When deleting a BUD it is now possible to check if the
deletion went without errors. This required a new
delete_ routine which now passes the BUD instead of the
BUD_ type (to be able to access the error field).

It is now possible to set the error message using any
integer precision.

Implemented read/write functions in:
- Dist1D-Array (needs to be re-checked)

Signed-off-by: Nick Papior <nickpapior@gmail.com>