bug: fixed OO compilation of all routines
Fixed all bugs currently existing when compiling with OO=1.
There were inconsistencies with the internal error value
and the function error (for retrieving the error message).
Now the internal variable has a suffix _ to disconnect it from
the procedure.

This is reflected in List/File/MP_Comm and Dist1D

Signed-off-by: Nick Papior <nickpapior@gmail.com>