build: fixed source construction, enabled building without FPP
Now one can build _only_ the sources and thus never
need to rely on a preprocessor.

When distributing the buds library we will pre-process all
sources in the, currently, 4 variants, and then one may do:

   make copy MPI=1

to copy all sources with MPI enabled into the current directory.
Then one may subsequently compile without the need of a good FPP.
This is an extra step for the user, but much less pain if they
use different compilers and/or are on an old system.

Worked around the need for a temporary file for preprocessing.
Now we use pipes to create the correct targets.

Signed-off-by: Nick Papior <nickpapior@gmail.com>