bug: fixed addition of custom routines for the finitestack, collections etc.
The logic when using the include statements was not consistent. Now one
can add routines and interfaces as one pleases for each of these extensions.

Another bug related to elemental size functions has been removed.
I think we should rename all size functions to something different.
Consider having an array of buds, calling size( ... ) on the array
will, by default, call size of each of the elements in the array,
if the overloaded function is elemental. However, what if we want the
size of the array: UBOUND - LBOUND + 1, tedious an unneccesary.
Hence, the functions should be renamed to something not intrinsic in
fortran.
TODO: rename size to different name.

Added fs_array_p for the FS_Array to retrieve an array from a finitestack.

Signed-off-by: Nick Papior <nickpapior@gmail.com>